ETL_project/requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
faker
python-dotenv
